+++ ariadne_codegen/client_generators/dependencies/async_base_client_open_telemetry.py
@@ -373,6 +373,12 @@ class AsyncBaseClientOpenTelemetry:
             **merged_kwargs,
         ) as websocket:
             await self._send_connection_init(websocket)
+            # wait for connection_ack from server
+            await self._handle_ws_message(
+                await websocket.recv(),
+                websocket,
+                expected_type=GraphQLTransportWSMessageType.CONNECTION_ACK,
+            )
             await self._send_subscribe(
                 websocket,
                 operation_id=operation_id,
@@ -414,7 +420,10 @@ class AsyncBaseClientOpenTelemetry:
         await websocket.send(json.dumps(payload))
 
     async def _handle_ws_message(
-        self, message: Data, websocket: WebSocketClientProtocol
+        self,
+        message: Data,
+        websocket: WebSocketClientProtocol,
+        expected_type: Optional[GraphQLTransportWSMessageType] = None,
     ) -> Optional[Dict[str, Any]]:
         try:
             message_dict = json.loads(message)
@@ -426,6 +435,11 @@ class AsyncBaseClientOpenTelemetry:
 
         if not type_ or type_ not in {t.value for t in GraphQLTransportWSMessageType}:
             raise GraphQLClientInvalidMessageFormat(message=message)
+
+        if expected_type and expected_type != type_:
+            raise GraphQLClientInvalidMessageFormat(
+                f"Invalid message received. Expected: {expected_type.value}"
+            )
 
         if type_ == GraphQLTransportWSMessageType.NEXT:
             if "data" not in payload:
@@ -563,6 +577,13 @@ class AsyncBaseClientOpenTelemetry:
                     root_span=root_span,
                     websocket=websocket,
                 )
+                # wait for connection_ack from server
+                await self._handle_ws_message_with_telemetry(
+                    root_span=root_span,
+                    message=await websocket.recv(),
+                    websocket=websocket,
+                    expected_type=GraphQLTransportWSMessageType.CONNECTION_ACK,
+                )
                 await self._send_subscribe_with_telemetry(
                     root_span=root_span,
                     websocket=websocket,
@@ -628,7 +649,11 @@ class AsyncBaseClientOpenTelemetry:
             )
 
     async def _handle_ws_message_with_telemetry(
-        self, root_span: Span, message: Data, websocket: WebSocketClientProtocol
+        self,
+        root_span: Span,
+        message: Data,
+        websocket: WebSocketClientProtocol,
+        expected_type: Optional[GraphQLTransportWSMessageType] = None,
     ) -> Optional[Dict[str, Any]]:
         with self.tracer.start_as_current_span(  # type: ignore
             "received message", context=set_span_in_context(root_span)
@@ -649,6 +674,11 @@ class AsyncBaseClientOpenTelemetry:
                 t.value for t in GraphQLTransportWSMessageType
             }:
                 raise GraphQLClientInvalidMessageFormat(message=message)
+
+            if expected_type and expected_type != type_:
+                raise GraphQLClientInvalidMessageFormat(
+                    f"Invalid message received. Expected: {expected_type.value}"
+                )
 
             if type_ == GraphQLTransportWSMessageType.NEXT:
                 if "data" not in payload: