+++ tests/client_generators/dependencies/test_websockets_open_telemetry.py
@@ -26,12 +26,19 @@ def mocked_websocket(mocked_ws_connect):
     websocket.__aiter__.return_value = [
         json.dumps({"type": "connection_ack"}),
     ]
+    websocket.recv.return_value = json.dumps({"type": "connection_ack"})
+    return websocket
+
+
+@pytest.fixture
+def mocked_faulty_websocket(mocked_ws_connect):
+    websocket = mocked_ws_connect.return_value.__aenter__.return_value
     return websocket
 
 
 @pytest.mark.asyncio
 async def test_execute_ws_creates_websocket_connection_with_correct_url(
-    mocked_ws_connect,
+    mocked_ws_connect, mocked_websocket  # pylint: disable=unused-argument
 ):
     async for _ in AsyncBaseClientOpenTelemetry(ws_url="ws://test_url").execute_ws(""):
         pass
@@ -42,7 +49,7 @@ async def test_execute_ws_creates_websocket_connection_with_correct_url(
 
 @pytest.mark.asyncio
 async def test_execute_ws_creates_websocket_connection_with_correct_subprotocol(
-    mocked_ws_connect,
+    mocked_ws_connect, mocked_websocket  # pylint: disable=unused-argument
 ):
     async for _ in AsyncBaseClientOpenTelemetry().execute_ws(""):
         pass
@@ -55,7 +62,7 @@ async def test_execute_ws_creates_websocket_connection_with_correct_subprotocol(
 
 @pytest.mark.asyncio
 async def test_execute_ws_creates_websocket_connection_with_correct_origin(
-    mocked_ws_connect,
+    mocked_ws_connect, mocked_websocket  # pylint: disable=unused-argument
 ):
     async for _ in AsyncBaseClientOpenTelemetry(ws_origin="test_origin").execute_ws(""):
         pass
@@ -66,7 +73,7 @@ async def test_execute_ws_creates_websocket_connection_with_correct_origin(
 
 @pytest.mark.asyncio
 async def test_execute_ws_creates_websocket_connection_with_correct_headers(
-    mocked_ws_connect,
+    mocked_ws_connect, mocked_websocket  # pylint: disable=unused-argument
 ):
     async for _ in AsyncBaseClientOpenTelemetry(
         ws_headers={"test_key": "test_value"}
@@ -82,7 +89,7 @@ async def test_execute_ws_creates_websocket_connection_with_correct_headers(
 @pytest.mark.asyncio
 @pytest.mark.parametrize("tracer", ["tracer name", None])
 async def test_execute_ws_creates_websocket_connection_with_passed_extra_headers(
-    mocked_ws_connect, tracer
+    mocked_ws_connect, mocked_websocket, tracer  # pylint: disable=unused-argument
 ):
     async for _ in AsyncBaseClientOpenTelemetry(
         ws_headers={"Client-A": "client_value_a", "Client-B": "client_value_b"},
@@ -103,7 +110,7 @@ async def test_execute_ws_creates_websocket_connection_with_passed_extra_headers
 @pytest.mark.asyncio
 @pytest.mark.parametrize("tracer", ["tracer name", None])
 async def test_execute_ws_creates_websocket_connection_with_passed_kwargs(
-    mocked_ws_connect, tracer
+    mocked_ws_connect, mocked_websocket, tracer  # pylint: disable=unused-argument
 ):
     async for _ in AsyncBaseClientOpenTelemetry(tracer=tracer).execute_ws(
         "", open_timeout=15, close_timeout=30
@@ -259,6 +266,19 @@ async def test_execute_ws_raises_graphql_multi_error_for_message_with_error_type
             pass
 
 
+@pytest.mark.asyncio
+async def test_execute_ws_raises_invalid_message_format_for_missing_ack_after_init(
+    mocked_faulty_websocket,
+):
+    mocked_faulty_websocket.recv.return_value = json.dumps(
+        {"type": "next", "payload": {"data": "test_data"}}
+    )
+
+    with pytest.raises(GraphQLClientInvalidMessageFormat):
+        async for _ in AsyncBaseClientOpenTelemetry().execute_ws(""):
+            pass
+
+
 @pytest.fixture
 def mocked_start_as_current_span(mocker):
     mocker_get_tracer = mocker.patch(
@@ -394,3 +414,22 @@ async def test_execute_ws_creates_span_for_received_error_message(
     mocked_start_as_current_span.assert_any_call("received message", context=ANY)
     with mocked_start_as_current_span.return_value as span:
         span.set_attribute.assert_any_call("type", "error")
+
+
+@pytest.mark.asyncio
+async def test_execute_ws_ws_creates_span_for_missing_ack_after_init(
+    mocked_faulty_websocket, mocked_start_as_current_span
+):
+    mocked_faulty_websocket.recv.return_value = json.dumps(
+        {"type": "next", "payload": {"data": "test_data"}}
+    )
+
+    client = AsyncBaseClientOpenTelemetry(ws_url="ws://test_url", tracer="tracker")
+
+    with pytest.raises(GraphQLClientInvalidMessageFormat):
+        async for _ in client.execute_ws(""):
+            pass
+
+    mocked_start_as_current_span.assert_any_call("received message", context=ANY)
+    with mocked_start_as_current_span.return_value as span:
+        span.set_attribute.assert_any_call("type", "next")