+++ tests/main/clients/extended_models/expected_client/fragments.py
@@ -20,3 +20,8 @@ class GetQueryAFragment(BaseModel):
 
 class GetQueryAFragmentQueryA(BaseModel, MixinA, CommonMixin):
     field_a: int = Field(alias="fieldA")
+
+
+FragmentA.model_rebuild()
+FragmentB.model_rebuild()
+GetQueryAFragment.model_rebuild()