+++ tests/main/clients/multiple_fragments/expected_client/fragments.py
@@ -40,3 +40,10 @@ class MinimalA(BaseModel):
 
 class MinimalAFieldB(MinimalB):
     pass
+
+
+CompleteA.model_rebuild()
+FullB.model_rebuild()
+FullA.model_rebuild()
+MinimalB.model_rebuild()
+MinimalA.model_rebuild()