+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed `graphql-transport-ws` protocol implementation not waiting for the `connection_ack` message on new connection.
 - Fixed `get_client_settings` mutating `config_dict` instance.
 - Added support to `graphqlschema` for saving schema as a GraphQL file.
+- Restored `model_rebuild` calls for top level fragment models.
 
 
 ## 0.11.0 (2023-12-05)